refactor: 메서드 분리, 패키지 정리
